refactor: FollowController, fix error for RevokeExistingTokens
